move UfsUtils to tachyon.underfs package
